+++ storybook/storybook-static/twig/precache-manifest.8d57ccf855ca62480905e708a5764569.js
@@ -1,0 +1,15 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "7490eb0987be34ac7af2ca6798350690",
+    "url": "iframe.html"
+  },
+  {
+    "url": "main.2a93ce6e9e86d64762a5.bundle.js"
+  },
+  {
+    "url": "runtime~main.2a93ce6e9e86d64762a5.bundle.js"
+  },
+  {
+    "url": "vendors~main.2a93ce6e9e86d64762a5.bundle.js"
+  }
+]);