+++ src/react-components/CHANGELOG.md
@@ -6,20 +6,31 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 
 ## PRERELEASE 
 ### Added 
+- `Dropdown` which controls `FormDropdown` and its corresponding `Label`
+- `Label` 
+- `HelperErrorText` 
+- `Pagination`
 - `EditionCard` component
 - `SearchResultItem` component that uses `EditionCard` 
 - `RN Header With Search` 
 - a story for `RN Section Title`, which is just a collection of styles
 
 ### Changed
+- added `iconModifiers` to `Button`
+- `buttonId` in `Button` is now required
+- refactored `FormDropdown` to not include its own `label`
+- `RN Header With Search` 
+- a story for `RN Section Title`, which is just a collection of styles
 - `Link` -> `BasicLink` for clarity
 - Bugfixes in `UnderlineLink` and `IconLink`
 - Added stories for `BasicLink`
 - `Link` missing URL error message
 - Consolidated `Heading` and `PageTitle`
+- `Heading` now only takes a single span
 - Added ResearchNow specific `SearchBar` error state 
 - A11Y changes for `SearchBar` and `HeaderWithSearch`
 - Added stories for AT-79, AT-264 and AT-3
+- `selectedOption` added to `FormDropdown` stores selected state
 
 ## [0.0.7] - 2020-01-03
 ### Added