+++ src/react-components/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease.  When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## PRERELEASE 
+## [0.0.8] - 2020-01-14 
 ### Added 
 - `Dropdown` which controls `FormDropdown` and its corresponding `Label`
 - `Label`